# Optimize attention weights multiplication with values