--- a/pkg/bitnet/internal/math/attention.go
+++ b/pkg/bitnet/internal/math/attention.go
@@ -59,7 +59,21 @@ func ScaledDotProductAttention(q, k, v *tensor.Tensor) *tensor.Tensor {
 				for j := 0; j < seqLen; j++ {
 					var sum float32
 					// Compute dot product between q[i] and k[j]
-					for d := 0; d < headDim; d++ {
+					// Process 4 elements at a time for better cache utilization
+					d := 0
+					for ; d+3 < headDim; d += 4 {
+						q0 := float32(q.Get(i, d))
+						q1 := float32(q.Get(i, d+1))
+						q2 := float32(q.Get(i, d+2))
+						q3 := float32(q.Get(i, d+3))
+						k0 := float32(k.Get(j, d))
+						k1 := float32(k.Get(j, d+1))
+						k2 := float32(k.Get(j, d+2))
+						k3 := float32(k.Get(j, d+3))
+						sum += q0*k0 + q1*k1 + q2*k2 + q3*k3
+					}
+					// Process remaining elements
+					for ; d < headDim; d++ {
 						sum += float32(q.Get(i, d)) * float32(k.Get(j, d))
 					}
 					// Scale by 1/sqrt(head_dim)
@@ -105,7 +119,7 @@ func ScaledDotProductAttention(q, k, v *tensor.Tensor) *tensor.Tensor {
 	}
 	wg.Wait()
 
-	// Compute weighted sum of values
+	// Compute weighted sum of values using higher precision for accumulation
 	output := tensor.NewTensor(seqLen, headDim)
 	for i := 0; i < seqLen; i += chunkSize {
 		wg.Add(1)
@@ -117,13 +131,35 @@ func ScaledDotProductAttention(q, k, v *tensor.Tensor) *tensor.Tensor {
 			}
 
 			for i := start; i < end; i++ {
-				for d := 0; d < headDim; d++ {
+				// Process 4 dimensions at a time for better cache utilization
+				d := 0
+				for ; d+3 < headDim; d += 4 {
+					var sum0, sum1, sum2, sum3 float32
+					// Accumulate in higher precision (float32)
+					for j := 0; j < seqLen; j++ {
+						w := weights[i][j]
+						v0 := float32(v.Get(j, d))
+						v1 := float32(v.Get(j, d+1))
+						v2 := float32(v.Get(j, d+2))
+						v3 := float32(v.Get(j, d+3))
+						sum0 += w * v0
+						sum1 += w * v1
+						sum2 += w * v2
+						sum3 += w * v3
+					}
+					// Clamp to int8 range and convert back to int8
+					output.Set(int8(min(max(int32(math.Round(float64(sum0))), -128), 127)), i, d)
+					output.Set(int8(min(max(int32(math.Round(float64(sum1))), -128), 127)), i, d+1)
+					output.Set(int8(min(max(int32(math.Round(float64(sum2))), -128), 127)), i, d+2)
+					output.Set(int8(min(max(int32(math.Round(float64(sum3))), -128), 127)), i, d+3)
+				}
+				// Process remaining dimensions
+				for ; d < headDim; d++ {
 					var sum float32
 					for j := 0; j < seqLen; j++ {
 						sum += weights[i][j] * float32(v.Get(j, d))
 					}
-					// Convert back to int8
-					output.Set(int8(math.Round(float64(sum))), i, d)
+					output.Set(int8(min(max(int32(math.Round(float64(sum))), -128), 127)), i, d)
 				}
 			}
 		}(i)
@@ -131,4 +167,20 @@ func ScaledDotProductAttention(q, k, v *tensor.Tensor) *tensor.Tensor {
 	wg.Wait()
 
 	return output
+}
+
+// min returns the minimum of two int32 values
+func min(a, b int32) int32 {
+	if a < b {
+		return a
+	}
+	return b
+}
+
+// max returns the maximum of two int32 values
+func max(a, b int32) int32 {
+	if a > b {
+		return a
+	}
+	return b
 }